requirements.txt
tensorflow
flask
flask-cors
pillow
numpy
gunicorn
kaggle
